refactor: 로그인 필요한 테스트 stub 요청으로 대체, 인증 로직 변경에 따른 테스트 수정
