+++ backend/main-service/src/test/java/kkakka/mainservice/AcceptanceTest.java
@@ -1,10 +1,23 @@
 package kkakka.mainservice;
 
+import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
+import static com.github.tomakehurst.wiremock.client.WireMock.post;
+import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
+import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
+import static kkakka.mainservice.fixture.TestDataLoader.MEMBER;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import io.restassured.RestAssured;
+import kkakka.mainservice.member.auth.application.dto.UserProfile;
 import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.annotation.DirtiesContext.ClassMode;
 import org.springframework.test.context.ActiveProfiles;
@@ -12,14 +25,35 @@ import org.springframework.test.context.ActiveProfiles;
 @ActiveProfiles("test")
 @DirtiesContext(classMode = ClassMode.AFTER_CLASS)
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+@AutoConfigureWireMock(port = 9001)
 public abstract class AcceptanceTest {
 
     @LocalServerPort
     int port;
+    @Autowired
+    private ObjectMapper objectMapper;
 
     @BeforeEach
-    public void setUp() {
+    public void setUp() throws JsonProcessingException {
         RestAssured.port = port;
+
+        stubFor(
+                post(urlEqualTo("/api/auth"))
+                        .willReturn(aResponse()
+                                .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
+                                .withStatus(HttpStatus.OK.value())
+                                .withBody(
+                                        objectMapper.writeValueAsString(
+                                                new UserProfile(
+                                                        MEMBER.getProvider(),
+                                                        MEMBER.getName(),
+                                                        MEMBER.getEmail(),
+                                                        MEMBER.getAgeGroup(),
+                                                        MEMBER.getPhone()
+                                                )
+                                        )
+                                )
+                        ));
     }
 }
