+++ backend/main-service/src/test/java/kkakka/mainservice/member/domain/MemberRepositoryTest.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import kkakka.mainservice.TestContext;
 import kkakka.mainservice.member.member.domain.Member;
-import kkakka.mainservice.member.member.domain.ProviderName;
 import kkakka.mainservice.member.member.domain.Provider;
+import kkakka.mainservice.member.member.domain.ProviderName;
 import kkakka.mainservice.member.member.domain.repository.MemberRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -25,7 +25,7 @@ public class MemberRepositoryTest extends TestContext {
         // given
         final Member member = memberRepository.save(
                 Member.create(
-                        Provider.create(TEST_MEMBER_01.getUserProfile().getId(),
+                        Provider.create(TEST_MEMBER_01.getProviderId(),
                                 ProviderName.TEST),
                         TEST_MEMBER_01.getName(),
                         TEST_MEMBER_01.getEmail(),
@@ -49,7 +49,7 @@ public class MemberRepositoryTest extends TestContext {
         // given
         final Member member = memberRepository.save(
                 Member.create(
-                        Provider.create(TEST_MEMBER_01.getUserProfile().getId(),
+                        Provider.create(TEST_MEMBER_01.getProviderId(),
                                 ProviderName.TEST),
                         TEST_MEMBER_01.getName(),
                         TEST_MEMBER_01.getEmail(),
@@ -73,7 +73,7 @@ public class MemberRepositoryTest extends TestContext {
         // given
         final Member member = memberRepository.save(
                 Member.create(
-                        Provider.create(TEST_MEMBER_01.getUserProfile().getId(),
+                        Provider.create(TEST_MEMBER_01.getProviderId(),
                                 ProviderName.TEST),
                         TEST_MEMBER_01.getName(),
                         TEST_MEMBER_01.getEmail(),
@@ -98,7 +98,7 @@ public class MemberRepositoryTest extends TestContext {
         // given
         final Member member = memberRepository.save(
                 Member.create(
-                        Provider.create(TEST_MEMBER_01.getUserProfile().getId(),
+                        Provider.create(TEST_MEMBER_01.getProviderId(),
                                 ProviderName.TEST),
                         TEST_MEMBER_01.getName(),
                         TEST_MEMBER_01.getEmail(),
@@ -122,7 +122,7 @@ public class MemberRepositoryTest extends TestContext {
         // given
         final Member member = memberRepository.save(
                 Member.create(
-                        Provider.create(TEST_MEMBER_01.getUserProfile().getId(),
+                        Provider.create(TEST_MEMBER_01.getProviderId(),
                                 ProviderName.TEST),
                         TEST_MEMBER_01.getName(),
                         TEST_MEMBER_01.getEmail(),
@@ -147,7 +147,7 @@ public class MemberRepositoryTest extends TestContext {
         // given
         final Member member = memberRepository.save(
                 Member.create(
-                        Provider.create(TEST_MEMBER_01.getUserProfile().getId(),
+                        Provider.create(TEST_MEMBER_01.getProviderId(),
                                 ProviderName.TEST),
                         TEST_MEMBER_01.getName(),
                         TEST_MEMBER_01.getEmail(),
